Fix padding logic in convolve_grayscale function for accurate image padding

+++ math/convolutions_and_pooling/3-convolve_grayscale.py
@@ -43,15 +43,15 @@ def convolve_grayscale(images, kernel, padding='same', stride=(1, 1)):
     # Pad images symmetrically
     padded_images = np.pad(
         images,
-        pad_width=((0, 0), (pad_h, pad_h), (pad_w, pad_w)),
+        pad_width=((0, 0), (int(pad_h), int(pad_h)), (int(pad_w), int(pad_w))),
         mode='constant'
     )
 
     m, h, w = padded_images.shape
 
     # Compute output dimensions
-    conv_h = int((h + 2 * pad_h - kh) // sh + 1)
-    conv_w = int((w + 2 * pad_w - kw) // sw + 1)
+    conv_h = (h + 2 * pad_h - kh) // sh + 1
+    conv_w = (w + 2 * pad_w - kw) // sw + 1
     output = np.zeros((m, conv_h, conv_w))
 
     # Perform convolution
